create: - added film-details-presenter and added reRender method of filmDetailsControlsComponent
