+++ src/view/film-details/film-details-top-container-view.js
@@ -1,0 +1,13 @@
+import ComponentView from '../component-view';
+
+const createTopContainer = () => ('<div class="film-details__top-container"></div>');
+
+export default class FilmDetailsTopContainerView extends ComponentView {
+  constructor() {
+    super();
+  }
+
+  get template() {
+    return createTopContainer();
+  }
+}
